refactor: partially ended the enter_screen nightmare
